+++ .prettierrc.js
@@ -1,8 +1,8 @@
 module.exports = {
-    "tabWidth": 4,
-    "useTabs": false,
-    "semi": false,
-    "singleQuote": true,
-    "trailingComma": "es5",
-    "endOfLine": "lf"
+    tabWidth: 4,
+    useTabs: false,
+    semi: false,
+    singleQuote: true,
+    trailingComma: 'es5',
+    endOfLine: 'lf',
 }